arreglo de erroes del lint
